Fixing fragment mess-up
