+++ app/app/src/main/java/me/bickositieff/raspio/ui/playlist/PlaylistViewModel.kt
@@ -1,10 +1,10 @@
-package me.bickositieff.raspio.ui.home
+package me.bickositieff.raspio.ui.playlist
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class PlaylistViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is home Fragment"
